refactor: index.ts => useComments.ts
